Simplify telemetry
