+++ tsconfig.dist.json
@@ -14,7 +14,8 @@
     "allowJs": true,
     "checkJs": true,
     "strict": true,
-    "moduleResolution": "Node"
+    "moduleResolution": "Node",
+    "resolveJsonModule": true
   },
   "include": ["src/*.d.ts", "src/*.ts", "src/*.js", "src/*.svelte"],
   "references": [{ "path": "./tsconfig.node.json" }],
